Inital courses querys
